backend/requirements.txt
annotated-types==0.7.0
anyascii==0.3.3
appnope==0.1.4
asgiref==3.11.0
asttokens==3.0.1
babel==2.17.0
beautifulsoup4==4.14.3
boto3==1.42.21
botocore==1.42.21
certifi==2026.1.4
charset-normalizer==3.4.4
comm==0.2.3
contextlib2==21.6.0
contourpy==1.3.3
cycler==0.12.1
debugpy==1.8.19
decorator==5.2.1
defusedxml==0.7.1
Django==6.0
django-cors-headers==4.9.0
django-countries==8.2.0
django-filter==25.2
django-modelcluster==6.4.1
django-money==3.5.4
django-ninja==1.4.3
django-ninja-extra==0.30.1
django-permissionedforms==0.1
django-phonenumber-field==8.4.0
django-schema-viewer==0.5.3
django-storages==1.14.6
django-stubs-ext==5.2.8
django-taggit==6.1.0
django-tasks==0.9.0
django-treebeard==4.8.0
djangorestframework==3.16.1
dnspython==2.8.0
draftjs_exporter==5.2.0
email-validator==2.3.0
et_xmlfile==2.0.0
exceptiongroup==1.3.1
executing==2.2.1
filetype==1.2.0
fonttools==4.61.1
idna==3.11
importlib_metadata==8.7.1
injector==0.22.0
ipython==9.8.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jmespath==1.0.1
kiwisolver==1.4.9
laces==0.1.2
matplotlib==3.10.8
matplotlib-inline==0.2.1
modelsearch==1.1.1
nest-asyncio==1.6.0
numpy==2.3.2
openpyxl==3.1.5
packaging==25.0
pandas==2.3.2
parso==0.8.5
pexpect==4.9.0
phonenumbers==9.0.21
pillow==12.1.0
pillow_heif==1.1.1
prompt_toolkit==3.0.52
psutil==7.2.1
psycopg2-binary==2.9.11
ptyprocess==0.7.0
pure_eval==0.2.3
py-moneyed==3.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pyparsing==3.3.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
pyzmq==27.1.0
requests==2.32.5
s3transfer==0.16.0
setuptools==80.9.0
six==1.17.0
soupsieve==2.8.1
SQLAlchemy==2.0.43
sqlparse==0.5.5
stack-data==0.6.3
telepath==0.3.1
tornado==6.5.4
traitlets==5.14.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.6.2
uWSGI==2.0.31
wagtail==7.2.1
wagtail-headless-preview
wagtail-ninja==0.1.17
wcwidth==0.2.14
Willow==1.12.0
wordcloud==1.9.5
zipp==3.23.0
